Adding sysctl setting only for dockerswarm

+++ e2e/e2e_tests/cluster_test_base.py
@@ -137,21 +137,20 @@ class TestClusterBase:
             )
             self.ssh_obj.make_directory(node=node, dir_name=self.docker_logs_path)
 
-            containers = self.ssh_obj.get_running_containers(node_ip=node)
-            self.container_nodes[node] = containers
-            self.ssh_obj.check_tmux_installed(node_ip=node)
-            self.ssh_obj.exec_command(node=node,
-                                      command="sudo tmux kill-server")
             if not self.k8s_test:
+                containers = self.ssh_obj.get_running_containers(node_ip=node)
+                self.container_nodes[node] = containers
+                self.ssh_obj.check_tmux_installed(node_ip=node)
+                self.ssh_obj.exec_command(node=node,
+                                        command="sudo tmux kill-server")
                 self.ssh_obj.start_docker_logging(node_ip=node,
                                                 containers=containers,
                                                 log_dir=self.docker_logs_path,
                                                 test_name=self.test_name
                                                 )
-            self.ssh_obj.start_tcpdump_logging(node_ip=node, log_dir=self.docker_logs_path)
-            self.ssh_obj.start_netstat_dmesg_logging(node_ip=node,
-                                                     log_dir=self.docker_logs_path)
-            if not self.k8s_test:
+                self.ssh_obj.start_tcpdump_logging(node_ip=node, log_dir=self.docker_logs_path)
+                self.ssh_obj.start_netstat_dmesg_logging(node_ip=node,
+                                                        log_dir=self.docker_logs_path)
                 self.ssh_obj.reset_iptables_in_spdk(node_ip=node)
         
         self.ssh_obj.delete_old_folders(
@@ -162,14 +161,15 @@ class TestClusterBase:
 
         self.ssh_obj.make_directory(node=self.fio_node, dir_name=self.docker_logs_path)
 
-        self.ssh_obj.check_tmux_installed(node_ip=self.fio_node)
+        if not self.k8s_test:
+            self.ssh_obj.check_tmux_installed(node_ip=self.fio_node)
 
-        self.ssh_obj.exec_command(node=self.fio_node,
-                                  command="sudo tmux kill-server")
-        self.ssh_obj.start_tcpdump_logging(node_ip=self.fio_node,
-                                           log_dir=self.docker_logs_path)
-        self.ssh_obj.start_netstat_dmesg_logging(node_ip=self.fio_node,
-                                                 log_dir=self.docker_logs_path)
+            self.ssh_obj.exec_command(node=self.fio_node,
+                                    command="sudo tmux kill-server")
+            self.ssh_obj.start_tcpdump_logging(node_ip=self.fio_node,
+                                            log_dir=self.docker_logs_path)
+            self.ssh_obj.start_netstat_dmesg_logging(node_ip=self.fio_node,
+                                                    log_dir=self.docker_logs_path)
 
         self.logger.info("Started log monitoring for all storage nodes.")
